cardinality query working

+++ packages/db/clibs/include/selva/hll.h
@@ -15,7 +15,7 @@ void hll_init(struct selva_string *hllss, uint8_t precision, bool is_sparse);
 SELVA_EXPORT
 void hll_add(struct selva_string *hllss, uint64_t element);
 SELVA_EXPORT
-uint32_t *hll_count(struct selva_string *hllss);
+uint8_t *hll_count(struct selva_string *hllss);
 SELVA_EXPORT
 struct selva_string hll_array_union(struct selva_string *hll_array, size_t count);
